refactor: View 컴포넌트를 `components/ui/atoms`로 이동
