+++ app/(tabs)/(onboard)/agreement.tsx
@@ -1,6 +1,6 @@
 import {StyleSheet} from 'react-native';
 import {TextSize} from '@/enums/TextSize';
-import View from '@/components/View';
+import { View } from '@/components/ui/atoms'
 import {Button, StyledText} from '@/components/ui/atoms';
 import Checkbox from '@/components/Checkbox';
 import useAgreementStore from '@/stores/agreementStore';
