+++ app/(tabs)/(onboard)/category.tsx
@@ -1,6 +1,6 @@
 import { Keyboard, Pressable, StyleSheet } from 'react-native';
 import { TextSize } from '@/enums/TextSize';
-import View from '@/components/View';
+import { View } from '@/components/ui/atoms'
 import {Button, StyledText} from '@/components/ui/atoms';
 import { router } from 'expo-router';
 import Category from '@/components/Category';
