+++ app/(tabs)/(onboard)/signup.tsx
@@ -1,6 +1,6 @@
 import {Keyboard, Pressable, StyleSheet, useColorScheme} from 'react-native';
 import { TextSize } from '@/enums/TextSize';
-import View from '@/components/View';
+import { View } from '@/components/ui/atoms'
 import {Button, StyledText} from '@/components/ui/atoms';
 import Textfield from '@/components/Textfield';
 import { useEffect, useState } from 'react';
