+++ components/ui/atoms/index.ts
@@ -1,2 +1,3 @@
 export { default as Button } from './Button'
 export { default as StyledText } from './Text'
+export { default as View } from './View'
